Implemented Undo functionality.
Stack implementation of the player and gameboard states. This means deep copying of the gameboard array on undo, and creation of new objects every turn. While it appears fast, This is likely a good implementation to address/redo If I find I'm not processing nodes fast enough with AI later on.